chore: Update to 1.5.3 sources
