--- src/Uno.UI/Microsoft/UI/Xaml/Controls/RatingControl/RatingControlStates.cs
@@ -1,16 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See LICENSE in the project root for license information.
-// MUX Reference RatingControl.h, tag winui3/release/1.4.2
-
-namespace Microsoft/* UWP don't rename */.UI.Xaml.Controls;
-
-internal enum RatingControlStates
-{
-	Disabled = 0,
-	Set = 1,
-	PointerOverSet = 2,
-	PointerOverPlaceholder = 3, // Also functions as the pointer over unset state at the moment
-	Placeholder = 4,
-	Unset = 5,
-	Null = 6
-}
