--- src/Uno.UI/Microsoft/UI/Xaml/Controls/RatingControl/RatingInfoType.cs
@@ -1,12 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See LICENSE in the project root for license information.
-// MUX Reference RatingControl.h, tag winui3/release/1.4.2
-
-namespace Microsoft/* UWP don't rename */.UI.Xaml.Controls;
-
-internal enum RatingInfoType
-{
-	None,
-	Font,
-	Image
-}
