+++ src/Uno.UI/Microsoft/UI/Xaml/Controls/RatingControl/RatingItemImageInfo.cs
@@ -32,7 +32,11 @@ public partial class RatingItemImageInfo : RatingItemInfo
 	/// Identifies the DisabledImage dependency property.
 	/// </summary>
 	public static DependencyProperty DisabledImageProperty { get; } =
-		DependencyProperty.Register(nameof(DisabledImage), typeof(ImageSource), typeof(RatingItemImageInfo), new FrameworkPropertyMetadata(null));
+		DependencyProperty.Register(
+			nameof(DisabledImage),
+			typeof(ImageSource),
+			typeof(RatingItemImageInfo),
+			new FrameworkPropertyMetadata(null));
 
 	/// <summary>
 	/// Gets or sets an image that represents a rating element that has been set by the user.
@@ -47,7 +51,11 @@ public partial class RatingItemImageInfo : RatingItemInfo
 	/// Identifies the Image dependency property.
 	/// </summary>
 	public static DependencyProperty ImageProperty { get; } =
-		DependencyProperty.Register(nameof(Image), typeof(ImageSource), typeof(RatingItemImageInfo), new FrameworkPropertyMetadata(null));
+		DependencyProperty.Register(
+			nameof(Image),
+			typeof(ImageSource),
+			typeof(RatingItemImageInfo),
+			new FrameworkPropertyMetadata(null));
 
 	/// <summary>
 	/// Gets or sets an image that represents a rating element that is showing a placeholder value.
@@ -62,7 +70,11 @@ public partial class RatingItemImageInfo : RatingItemInfo
 	/// Identifies the PlaceholderImage dependency property.
 	/// </summary>
 	public static DependencyProperty PlaceholderImageProperty { get; } =
-		DependencyProperty.Register(nameof(PlaceholderImage), typeof(ImageSource), typeof(RatingItemImageInfo), new FrameworkPropertyMetadata(null));
+		DependencyProperty.Register(
+			nameof(PlaceholderImage),
+			typeof(ImageSource),
+			typeof(RatingItemImageInfo),
+			new FrameworkPropertyMetadata(null));
 
 	/// <summary>
 	/// Gets or sets an image that represents a rating element that has the pointer over it.
@@ -77,7 +89,11 @@ public partial class RatingItemImageInfo : RatingItemInfo
 	/// Identifies the PointerOverImage dependency property.
 	/// </summary>
 	public static DependencyProperty PointerOverImageProperty { get; } =
-		DependencyProperty.Register(nameof(PointerOverImage), typeof(ImageSource), typeof(RatingItemImageInfo), new FrameworkPropertyMetadata(null));
+		DependencyProperty.Register(
+			nameof(PointerOverImage),
+			typeof(ImageSource),
+			typeof(RatingItemImageInfo),
+			new FrameworkPropertyMetadata(null));
 
 	/// <summary>
 	/// Gets or sets an image that represents a rating element showing a placeholder value with the pointer over it.
@@ -92,7 +108,11 @@ public partial class RatingItemImageInfo : RatingItemInfo
 	/// Identifies the PointerOverPlaceholderImage dependency property.
 	/// </summary>
 	public static DependencyProperty PointerOverPlaceholderImageProperty { get; } =
-		DependencyProperty.Register(nameof(PointerOverPlaceholderImage), typeof(ImageSource), typeof(RatingItemImageInfo), new FrameworkPropertyMetadata(null));
+		DependencyProperty.Register(
+			nameof(PointerOverPlaceholderImage),
+			typeof(ImageSource),
+			typeof(RatingItemImageInfo),
+			new FrameworkPropertyMetadata(null));
 
 	/// <summary>
 	/// Gets or sets an image that represents a rating element that has not been set.
@@ -107,5 +127,9 @@ public partial class RatingItemImageInfo : RatingItemInfo
 	/// Identifies the UnsetImage dependency property.
 	/// </summary>
 	public static DependencyProperty UnsetImageProperty { get; } =
-		DependencyProperty.Register(nameof(UnsetImage), typeof(ImageSource), typeof(RatingItemImageInfo), new FrameworkPropertyMetadata(null));
+		DependencyProperty.Register(
+			nameof(UnsetImage),
+			typeof(ImageSource),
+			typeof(RatingItemImageInfo),
+			new FrameworkPropertyMetadata(null));
 }
